+++ src/tests/BasePluginTests.ts
@@ -18,42 +18,34 @@ process.env.PLUGIN_WORKER_ID = PLUGIN_WORKER_ID;
 describe('Plugin Status API Tests', function () {
   class MyFakePlugin extends core.BasePlugin {}
 
-  it('should return plugin status (200) if the plugin is OK', function (done) {
+  it('should return plugin status (200) if the plugin is OK', async function () {
     const plugin = new MyFakePlugin(false);
     const runner = new core.TestingPluginRunner(plugin);
 
-    void request(runner.plugin.app)
-      .get('/v1/status')
-      .end(function (err, res) {
-        expect(res.status).to.equal(200);
-        done();
-      });
+    const res = await request(runner.plugin.app).get('/v1/status');
+    expect(res.status).to.equal(200);
   });
 });
 
 describe('Plugin Metadata API Tests', function () {
   class MyFakePlugin extends core.BasePlugin {}
 
-  it('should return metadata', function (done) {
+  it('should return metadata', async function () {
     const plugin = new MyFakePlugin(false);
     const runner = new core.TestingPluginRunner(plugin);
 
-    void request(runner.plugin.app)
-      .get('/v1/metadata')
-      .end(function (err, res) {
-        expect(res.status).to.equal(200);
-        expect(JSON.stringify(res.body)).to.equal(
-          JSON.stringify({ runtime: 'node', runtime_version: process.version, dependencies: {} }),
-        );
-        done();
-      });
+    const res = await request(runner.plugin.app).get('/v1/metadata');
+    expect(res.status).to.equal(200);
+    expect(JSON.stringify(res.body)).to.equal(
+      JSON.stringify({ runtime: 'node', runtime_version: process.version, dependencies: {} }),
+    );
   });
 });
 
 describe('Plugin log level API tests', function () {
   class MyFakePlugin extends core.BasePlugin {}
 
-  it('Log Level update should return 200', function (done) {
+  it('Log Level update should return 200', async function () {
     const plugin = new MyFakePlugin(false);
     const runner = new core.TestingPluginRunner(plugin);
 
@@ -61,16 +53,11 @@ describe('Plugin log level API tests', function () {
       level: 'debug',
     };
 
-    void request(runner.plugin.app)
-      .put('/v1/log_level')
-      .send(requestBody)
-      .end(function (err, res) {
-        expect(res.status).to.equal(200);
-        done();
-      });
+    const res = await request(runner.plugin.app).put('/v1/log_level').send(requestBody);
+    expect(res.status).to.equal(200);
   });
 
-  it('Malformed Log level update should return 400', function (done) {
+  it('Malformed Log level update should return 400', async function () {
     const plugin = new MyFakePlugin(false);
     const runner = new core.TestingPluginRunner(plugin);
 
@@ -79,16 +66,11 @@ describe('Plugin log level API tests', function () {
       hector: 'debug',
     };
 
-    void request(runner.plugin.app)
-      .put('/v1/log_level')
-      .send(requestBody)
-      .end(function (err, res) {
-        expect(res.status).to.equal(400);
-        done();
-      });
+    const res = await request(runner.plugin.app).put('/v1/log_level').send(requestBody);
+    expect(res.status).to.equal(400);
   });
 
-  it('Should return WARN when getting Log Level', function (done) {
+  it('Should return WARN when getting Log Level', async function () {
     const plugin = new MyFakePlugin(false);
     const runner = new core.TestingPluginRunner(plugin);
 
@@ -96,20 +78,12 @@ describe('Plugin log level API tests', function () {
       level: 'WARN',
     };
 
-    void request(runner.plugin.app)
-      .put('/v1/log_level')
-      .send(requestBody)
-      .end(function (err, res) {
-        expect(res.status).to.equal(200);
-      });
+    const res1 = await request(runner.plugin.app).put('/v1/log_level').send(requestBody);
+    expect(res1.status).to.equal(200);
 
-    void request(runner.plugin.app)
-      .get('/v1/log_level')
-      .end(function (err, res) {
-        expect(res.status).to.equal(200);
-        expect(res.body.level).to.equal(requestBody.level);
-        done();
-      });
+    const res2 = await request(runner.plugin.app).get('/v1/log_level');
+    expect(res2.status).to.equal(200);
+    expect(res2.body.level).to.equal(requestBody.level);
   });
 });
 
@@ -118,7 +92,7 @@ describe('Request Gateway helper API tests', function () {
 
   class MyFakePlugin extends core.BasePlugin {}
 
-  it('Check that uri is passed correctly', function (done) {
+  it('Check that url is passed correctly', async function () {
     const plugin = new MyFakePlugin(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
@@ -126,26 +100,24 @@ describe('Request Gateway helper API tests', function () {
     const fakeMethod = 'GET';
 
     // We try a call to the Gateway
-    void runner.plugin.requestGatewayHelper('GET', fakeUri).then(() => {
+    await runner.plugin.requestGatewayHelper({ method: 'GET', url: fakeUri }).then(() => {
       expect(rpMockup.args[0][0].method).to.be.eq(fakeMethod);
-      expect(rpMockup.args[0][0].uri).to.be.eq(fakeUri);
-      done();
+      expect(rpMockup.args[0][0].url).to.be.eq(fakeUri);
     });
   });
 
-  it('Authentification token should be passed from values passed in the env', function (done) {
+  it('Authentification token should be passed from values passed in the env', async function () {
     const plugin = new MyFakePlugin(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
     // We try a call to the Gateway
-    void runner.plugin.requestGatewayHelper('GET', '/v1/easter_eggs/').then(() => {
-      expect(rpMockup.args[1][0].auth.pass).to.be.eq(PLUGIN_AUTHENTICATION_TOKEN);
-      expect(rpMockup.args[1][0].auth.user).to.be.eq(PLUGIN_WORKER_ID);
-      done();
+    await runner.plugin.requestGatewayHelper({ method: 'GET', url: '/v1/easter_eggs/' }).then(() => {
+      expect(rpMockup.args[1][0].headers.pass).to.be.eq(PLUGIN_AUTHENTICATION_TOKEN);
+      expect(rpMockup.args[1][0].headers.user).to.be.eq(PLUGIN_WORKER_ID);
     });
   });
 
-  it('Check that body is passed correctly when set', function (done) {
+  it('Check that body is passed correctly when set', async function () {
     const plugin = new MyFakePlugin(false);
     const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
@@ -154,11 +126,10 @@ describe('Request Gateway helper API tests', function () {
     const fakeBody = { sucess: true };
 
     // We try a call to the Gateway
-    void runner.plugin.requestGatewayHelper('GET', fakeUri, fakeBody).then(() => {
+    await runner.plugin.requestGatewayHelper({ method: 'GET', url: fakeUri, body: fakeBody }).then(() => {
       expect(rpMockup.args[2][0].method).to.be.eq(fakeMethod);
-      expect(rpMockup.args[2][0].uri).to.be.eq(fakeUri);
-      expect(rpMockup.args[2][0].body).to.be.eq(fakeBody);
-      done();
+      expect(rpMockup.args[2][0].url).to.be.eq(fakeUri);
+      expect(rpMockup.args[2][0].json).to.be.eq(fakeBody);
     });
   });
 });
@@ -173,36 +144,34 @@ describe('Data File helper Tests', function () {
   const plugin = new MyFakePlugin(false);
   const runner = new core.TestingPluginRunner(plugin, rpMockup);
 
-  it('DataFile: Should call the proper gateway URL', function (done) {
+  it('DataFile: Should call the proper gateway URL', async function () {
     const dataFileGatewayURI = '/v1/data_file/data';
     const method = 'GET';
     const fakeDataFileURI = 'mics://fake_dir/fake_file';
 
     // We try a call to the Gateway
-    void runner.plugin.fetchDataFile(fakeDataFileURI).then((file) => {
+    await runner.plugin.fetchDataFile(fakeDataFileURI).then((file) => {
       expect(rpMockup.args[0][0].method).to.be.eq(method);
-      expect(rpMockup.args[0][0].uri).to.be.eq(
+      expect(rpMockup.args[0][0].url).to.be.eq(
         `http://${runner.plugin.gatewayHost}:${runner.plugin.gatewayPort}${dataFileGatewayURI}`,
       );
-      expect(rpMockup.args[0][0].qs['uri']).to.be.eq(fakeDataFileURI);
+      expect(rpMockup.args[0][0].searchParams['uri']).to.be.eq(fakeDataFileURI);
       expect(file).to.be.eq(fakeDataFile);
-      done();
     });
   });
 
-  it('ConfigurationFile: Should call the proper gateway URL', function (done) {
+  it('ConfigurationFile: Should call the proper gateway URL', async function () {
     const confFileName = 'toto';
     const method = 'GET';
     const confFileGatewayURI = `/v1/configuration/technical_name=${confFileName}`;
 
     // We try a call to the Gateway
-    void runner.plugin.fetchConfigurationFile(confFileName).then((file) => {
+    await runner.plugin.fetchConfigurationFile(confFileName).then((file) => {
       expect(rpMockup.args[1][0].method).to.be.eq(method);
-      expect(rpMockup.args[1][0].uri).to.be.eq(
+      expect(rpMockup.args[1][0].url).to.be.eq(
         `http://${runner.plugin.gatewayHost}:${runner.plugin.gatewayPort}${confFileGatewayURI}`,
       );
       expect(file).to.be.eq(fakeDataFile);
-      done();
     });
   });
 });
@@ -210,13 +179,12 @@ describe('Data File helper Tests', function () {
 describe('Instance Context Expiration Tests', function () {
   class MyFakePlugin extends core.BasePlugin {}
 
-  it('InstanceContextExpiration: Check Instance Context variability: should be less than 10%', function (done) {
+  it('InstanceContextExpiration: Check Instance Context variability: should be less than 10%', function () {
     const plugin = new MyFakePlugin(false);
 
     const refreshInterval = plugin.getInstanceContextCacheExpiration();
 
     expect(refreshInterval).to.be.gte(plugin.INSTANCE_CONTEXT_CACHE_EXPIRATION);
     expect(refreshInterval).to.be.lte(plugin.INSTANCE_CONTEXT_CACHE_EXPIRATION * 1.1);
-    done();
   });
 });